🛠️ refactor|컴포넌트 내 api 요청상위 페이지로 이동, 로딩 gif 수정
