+++ FC.gg/fc_app/src/css/MatchResult.css
@@ -1,3 +1,7 @@
+.MatchResultLoading{
+  align-self: center;
+}
+
 .MatchResultBackground {
   display: flex;
   flex-direction: column;
