+++ FC.gg/fc_app/src/pages/ResultPage.js
@@ -14,18 +14,22 @@ export default function Screen() {
   const [matchtype, setMatch] = useState(50);
   const [nickname, setNickname] = useState("");
   const [ouid, setOuid] = useState("");
+  const [matchdetail, setMatchdetail] = useState("");
+  const [loading, setLoading] = useState(true);
   
   const onChange = (e) => {
     e.preventDefault();
     setSearchText(e.target.value);
   };
   const buttonClick = () => {
+    setNickname("");
     if (!searchText) {
       alert("검색어를 입력해주세요!");
     } else {
       navigate(`./?input=${searchText}`);
       setNickname(searchText)
       setSearchText("");
+      setLoading(true)
     }
   };
   const keyPress = (e) => {
@@ -60,8 +64,52 @@ export default function Screen() {
       
     }  
     handleGetouid();
-  }, [input, nickname, matchtype, ouid])
+  }, [input, nickname, ouid])
   console.log("ouidtest:", ouid)
+
+  useEffect(()=>{
+    const getMatchdetail = async() => {
+      try {
+        const response1 = await axios.get('https://p0l0evybh6.execute-api.ap-northeast-2.amazonaws.com/dev/Getmatchid', {
+          params: {
+            ouid: ouid,
+            matchtype: matchtype,
+            offset: 0,
+            limit: 10
+          }
+        });
+        setMatch(matchtype);
+  
+        console.log(response1);
+  
+        const matchData = {};
+        console.log('matchid', response1.data);
+  
+        await Promise.all(response1.data.map(async (id) => {
+          console.log('id', id);
+          const response2 = await axios.get('https://p0l0evybh6.execute-api.ap-northeast-2.amazonaws.com/dev/Getmatchdetail', {
+            params: {
+              matchid: String(id),
+              // nickname: sessionStorage.getItem('nickname') 세션 스토리지에 닉네임 넣는 코드가 없어서 그냥 setNickname 있길레 nickname으로 대체했슴다
+              nickname: nickname
+            }
+          });
+          // 응답 데이터를 처리하거나 상태에 저장
+
+          console.log(response2.data);
+          matchData[id] = response2.data;
+        }));
+      console.log(matchData);
+      setMatchdetail(matchData);
+      console.log("matchdetail:",matchdetail);
+      setLoading(false);
+      } catch (error) {
+        console.error('Error:', error);
+      }
+    }
+    getMatchdetail();
+  }, [ouid, matchtype, nickname])
+
   return (
     <div className="screen">
       <div className="div">
@@ -83,7 +131,7 @@ export default function Screen() {
           <div className="MatchTypeText">감독모드</div>
           <div className="MatchTypeText">클래식 1vs1</div>
         </div>
-        <MatchResult ouid={ouid} input={input} />
+        <MatchResult nickname={nickname} matchdetail={matchdetail} loading={loading} />
       </div>
     </div>
   );
